+++ appseed/ca/sockets/bsd/internal/SSLInitializer.h
@@ -53,7 +53,7 @@ namespace sockets
       static void SSL_locking_function(int32_t mode, int32_t n, const char *file, int32_t line);
 
       /** Return thread id. */
-      static uint32_t long SSL_id_function();
+      static uint32_t SSL_id_function();
 
       BIO *bio_err;
